+++ bin/run_electrometer.py
@@ -1,6 +1,0 @@
-#!/bin/env python
-import asyncio
-
-from lsst.ts.electrometer.csc import ElectrometerCsc
-
-asyncio.run(ElectrometerCsc.amain(index=True))